Fix broken travis deployment (#5)
* Add deploy stage to travis build
* Change conan package deployment format
* Fix CMakeLists.txt target and packaging
* Move '.ci' to '.travis' directory
* Change 'BUILD_SUFFIX' to 'CONAN_VERSION_SUFFIX'
* Fix conan version to 1.7.4

The most recent conan release breaks when installing on
travis's macos agents, due to the lack of the 'man' program.
1.7.4 does not have this issue, so use this version instead until
this is fixed.

* Remove Xcode7.3 and Xcode8 build agents

These two agents have been spuriously failing when installing
python3 and pip. Since these are older/more antiquated versions of
Xcode, support for them is less significant - and thus is being
phased out.

* Remove Xcode 8.1 and Xcode 8.2 travis CI images

The images are being remove from use due to a message on travis,
which states: "This job was configured to run on an OS X image that was
retired on November 28, 2017. It was routed to our Xcode 8.3 image."